% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Connect to external AMR database}
\usage{
connect_to_database(
  db_type = c("glass", "atlas", "ear", "other"),
  credentials = list(),
  config = list()
)
}
\arguments{
\item{db_type}{Type of database: "glass", "atlas", "ear", or other}

\item{credentials}{List containing authentication credentials}

\item{config}{Additional configuration parameters}
}
\value{
A connection object to the external database
}
\description{
Establishes a connection to an external AMR database using
provided credentials and configuration.
}
\examples{
\dontrun{
# Connect to WHO GLASS database
glass_credentials <- list(
  username = "myuser",
  password = "mypass",
  api_key = "my-api-key"
)

glass_conn <- connect_to_database(
  db_type = "glass",
  credentials = glass_credentials
)
}
}
