% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{get_data_model}
\alias{get_data_model}
\title{Define the standard data model for AMR data}
\usage{
get_data_model()
}
\value{
A list containing data model specifications
}
\description{
This function returns the standard data model structure for metero,
including metadata, human data, animal data, and environment data schemas.
}
\examples{
data_model <- get_data_model()
# View the metadata schema
data_model$metadata_schema
}
