% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{get_standard_amr_structure}
\alias{get_standard_amr_structure}
\title{Get standard AMR data structure}
\usage{
get_standard_amr_structure()
}
\value{
A data frame with variable names and descriptions
}
\description{
Returns the standard variable names and descriptions for AMR data in metero
}
