% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_amr}
\alias{human_amr}
\title{Human AMR Data}
\format{
A data frame with 591 rows and 16 variables:
\describe{
\item{study_id}{Unique identifier for each study}
\item{pathogen_name}{Full name of the pathogen}
\item{antibiotic_name}{Full name of the antibiotic}
\item{sample_count}{Number of samples tested}
\item{resistant_count}{Number of resistant isolates}
\item{resistance_rate}{Proportion of resistant isolates (0-1)}
\item{domain}{Domain (human, animal, or environment)}
\item{region}{Geographical region}
\item{author}{First author of the study}
\item{year}{Year of publication}
\item{country}{Country where the study was conducted}
\item{study_design}{Type of study design}
\item{setting}{Environment setting (e.g., rural, urban, hospital)}
\item{population_type}{Study population (e.g., adults, children)}
\item{total_sample_size}{Total sample size of the study}
\item{age_range}{Age range of study participants}
}
}
\source{
Meta-analysis of published studies on antimicrobial resistance in South Asia
}
\usage{
human_amr
}
\description{
A dataset containing antimicrobial resistance data from human populations.
This dataset contains information on community carriage of antimicrobial resistant
bacterial pathogens based on a systematic review and meta-analysis of studies
conducted in South Asian countries.
}
\keyword{datasets}
