% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{is_meteor_dataset}
\alias{is_meteor_dataset}
\title{Check if an object is a valid metero dataset}
\usage{
is_meteor_dataset(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if the object is a valid metero dataset, FALSE otherwise
}
\description{
Validates whether an object conforms to the metero dataset structure.
}
\examples{
dataset <- create_empty_dataset()
is_meteor_dataset(dataset)
}
