% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{perform_meta_analysis}
\alias{perform_meta_analysis}
\title{Internal function to perform meta-analysis on a data frame}
\usage{
perform_meta_analysis(data, method, measure, min_studies, weights, conf_level)
}
\arguments{
\item{data}{Data frame with AMR data}

\item{method}{Meta-analysis method}

\item{measure}{Effect measure type}

\item{min_studies}{Minimum studies required}

\item{weights}{Weighting approach}

\item{conf_level}{Confidence level}
}
\value{
List with meta-analysis results
}
\description{
Internal function to perform meta-analysis on a data frame
}
\keyword{internal}
