% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{prepare_meta_data}
\alias{prepare_meta_data}
\title{Prepare data for meta-regression analysis}
\usage{
prepare_meta_data(data, by)
}
\arguments{
\item{data}{AMR data frame}

\item{by}{Grouping variables}
}
\value{
List with meta-regression data
}
\description{
Prepare data for meta-regression analysis
}
\keyword{internal}
