% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{print.amr_meta_analysis}
\alias{print.amr_meta_analysis}
\title{Print method for AMR meta-analysis}
\usage{
\method{print}{amr_meta_analysis}(x, ...)
}
\arguments{
\item{x}{An amr_meta_analysis object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the meta-analysis object
}
\description{
Print method for AMR meta-analysis
}
