% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{print.meteor_dataset}
\alias{print.meteor_dataset}
\title{Print method for meteor dataset}
\usage{
\method{print}{meteor_dataset}(x, ...)
}
\arguments{
\item{x}{A meteor dataset}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the meteor dataset
}
\description{
Print method for meteor dataset
}
\examples{
\dontrun{
dataset <- create_empty_dataset()
print(dataset)
}
}
