% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{print.meteor_summary}
\alias{print.meteor_summary}
\title{Print method for meteor summary}
\usage{
\method{print}{meteor_summary}(x, ...)
}
\arguments{
\item{x}{A meteor dataset summary}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the summary object
}
\description{
Print method for meteor summary
}
